Change Button to SubmitButton and move Button to the right
